requirements.txt
curl_cffi
python-dotenv
2captcha-python